requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
python-dotenv
gunicorn
pyjwt
